+++ docs/CHANGELOG.rst
@@ -40,6 +40,7 @@ Fixed
 - Make parallel test suite worker threads clean up after initialization failures
 - Add mechanism to override the manager's control channel prefix from the
   environment
+- Make management commands respect the set verbosity level
 
 
 ==================